+++ lib/src/focus_debugger.dart
@@ -10,7 +10,14 @@ import 'package:flutter_focus_debugger/src/focus_widget_painter.dart';
 ///  - Visual bounds around the widget with primary focus.
 ///  - A translucent pane that shows the history of focus movement.
 class FocusDebugger extends StatelessWidget {
-  const FocusDebugger({super.key, this.isEnabled = true, required this.child});
+  const FocusDebugger({
+    super.key,
+    this.isEnabled = true,
+    this.showFocusBorder = true,
+    this.animateFocusBorderChanges = true,
+    this.showFocusHistory = true,
+    required this.child,
+  });
 
   /// Whether to show a focus debugging UI.
   ///
@@ -18,6 +25,19 @@ class FocusDebugger extends StatelessWidget {
   /// tree across debug and production builds, and simply toggle the behavior on
   /// and off.
   final bool isEnabled;
+
+  /// Whether to show a colorful border around the currently focused widget, when
+  /// `isEnabled` is `true`.
+  final bool showFocusBorder;
+
+  /// `true` for the focus bounds to animate from the previously focused widget to
+  /// the newly focused widget, or `false` to immediately update the focus border to
+  /// appear around the newly focused widget.
+  final bool animateFocusBorderChanges;
+
+  /// Whether to show a pane that lists the history of which widgets have had focus
+  /// since the start of the app, when `isEnabled` is `true`.
+  final bool showFocusHistory;
 
   /// The app widget tree (this widget is typically placed at the top of the widget
   /// tree).
@@ -33,8 +53,18 @@ class FocusDebugger extends StatelessWidget {
       textDirection: TextDirection.ltr,
       child: Stack(
         children: [
-          FocusWidgetPainter(child: child), //
-          Positioned(top: 0, right: 0, bottom: 0, child: FocusHistoryPane()),
+          if (showFocusBorder) //
+            FocusWidgetPainter(
+              animateBounds: animateFocusBorderChanges,
+              child: child,
+            ), //
+          if (showFocusHistory) //
+            Positioned(
+              top: 0,
+              right: 0,
+              bottom: 0,
+              child: FocusHistoryPane(),
+            ),
         ],
       ),
     );